requirements.txt
DB2mermaid==1.0.3
mysqlclient==2.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
